[REFACTOR] project 스프링 부트 전환
