+++ backend/src/main/java/com/animalfarm/backend/domain/token/TokenViewController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.animalfarm.backend.domain.token.dto.TokenListDTO;
+import com.animalfarm.backend.domain.token.dto.TokenSummaryDTO;
 import com.animalfarm.backend.global.http.ExternalApiClient;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -27,7 +27,7 @@ public class TokenViewController {
 
 	@GetMapping
 	public String tokenListPage(Model model) {
-		List<TokenListDTO> list = tokenService.selectAll();
+		List<TokenSummaryDTO> list = tokenService.selectAll();
 		if (!list.isEmpty()) {
 			System.out.println("Data Type: " + list.get(0).getClass().getName());
 		}
